test(openai): split responses provider tests
